Added FirebaseMessaging, timestamps are now double

+++ app/src/main/java/com/example/android/questtime/Question.java
@@ -7,14 +7,14 @@ package com.example.android.questtime;
 public class Question {
 
     private String text;
-    private long timestamp;
+    private double timestamp;
     private int points;
     private String id;
     private String category;
     private String correct;
     private String wrong[];
 
-    public Question(String text, long timestamp, int points, String id, String category) {
+    public Question(String text, double timestamp, int points, String id, String category) {
         this.text = text;
         this.timestamp = timestamp;
         this.points = points;
@@ -34,7 +34,7 @@ public class Question {
         return text;
     }
 
-    public long getTimestamp() {
+    public double getTimestamp() {
         return timestamp;
     }
